Create factory, utils packages
