+++ contracts/streamswap-factory/Cargo.toml
@@ -40,6 +40,7 @@ semver = { workspace = true }
 serde = { workspace = true }
 thiserror = { workspace = true }
 streamswap-types = { workspace = true }
+streamswap-utils = { workspace = true }
 
 [dev-dependencies]
 cw-multi-test = { workspace = true }
