+++ contracts/streamswap-factory/src/contract.rs
@@ -2,12 +2,12 @@ use cosmwasm_std::{
     entry_point, to_json_binary, Binary, Coin, CosmosMsg, Decimal, Deps, DepsMut, Env, MessageInfo,
     Response, StdResult, WasmMsg,
 };
-
-use crate::{
-    error::ContractError,
-    payment_checker::check_payment,
-    state::{Params, FREEZESTATE, LAST_STREAM_ID, PARAMS},
+use streamswap_types::factory::{
+    ContractError, CreateStreamMsg, ExecuteMsg, InstantiateMsg, Params, QueryMsg,
 };
+use streamswap_utils::payment_checker::check_payment;
+
+use crate::state::{FREEZESTATE, LAST_STREAM_ID, PARAMS};
 
 #[cfg_attr(not(feature = "library"), entry_point)]
 pub fn instantiate(
