+++ contracts/streamswap-factory/src/lib.rs
@@ -1,4 +1,2 @@
 pub mod contract;
-pub mod error;
-pub mod payment_checker;
 pub mod state;
