+++ packages/types/src/factory/mod.rs
@@ -1,3 +1,7 @@
+mod error;
 mod msg;
+mod types;
 
+pub use error::*;
 pub use msg::*;
+pub use types::*;
